% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_files_to_dataframe.R
\name{combine_files_to_dataframe}
\alias{combine_files_to_dataframe}
\title{Combine data files into dataframe}
\usage{
combine_files_to_dataframe(type)
}
\arguments{
\item{type}{A string naming the data file type to combine}
}
\value{
A dataframe of the combined data
}
\description{
Combine data files into dataframe
}
