% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(data_path = "data/msds_download")
}
\arguments{
\item{data_path}{Character string defining the path to the parent data folder}
}
\value{
A data frame of data
}
\description{
Get data
}
