% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dq.R
\name{get_dq}
\alias{get_dq}
\title{Get data quality data}
\usage{
get_dq(data_path = "data/msds_download")
}
\arguments{
\item{data_path}{Character string defining the path to the parent data folder}
}
\value{
A dataframe of data quality data
}
\description{
Get data quality data
}
