% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Get Measures}
\usage{
get_measures(data_path = "data/msds_download")
}
\arguments{
\item{data_path}{Character string defining the path to the parent data folder}
}
\value{
A dataframe of measures data
}
\description{
Get Measures
}
