% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msds_tidy_data.R
\name{msds_tidy_data}
\alias{msds_tidy_data}
\title{Combine and tidy data files}
\usage{
msds_tidy_data(data_path = "data/msds_download")
}
\arguments{
\item{data_path}{Character string defining the path to the parent data folder}
}
\value{
A data frame of data
}
\description{
Combine and tidy data files
}
