% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demo_dq.R
\name{plot_demo_dq}
\alias{plot_demo_dq}
\title{Make a comparative plot using data quality data
as an example of the data available}
\usage{
plot_demo_dq(dtf, focus_org_code)
}
\arguments{
\item{dtf}{A dataframe containing measures data}

\item{focus_org_code}{A character vector of a single org code to highlight in the plot}
}
\description{
Make a comparative plot using data quality data
as an example of the data available
}
