% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_measure_comparative.R
\name{plot_measure_comparative}
\alias{plot_measure_comparative}
\title{Make a comparative plot using measures data
ss an example of the data available}
\usage{
plot_measure_comparative(dtf, measure_name, focus_org_code)
}
\arguments{
\item{dtf}{A dataframe containing measures data}

\item{measure_name}{A character vector of the measure name to plot}

\item{focus_org_code}{A character vector of a single org code to highlight in the plot}
}
\description{
Make a comparative plot using measures data
ss an example of the data available
}
