% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msds_download_data.R
\name{msds_download_data}
\alias{msds_download_data}
\title{Check for, and download MSDS data}
\usage{
msds_download_data(
  destination = "data/downloaded",
  check_for_new_data = TRUE,
  force_redownload_all = FALSE
)
}
\arguments{
\item{destination}{Character. The filepath where downloads will be saved}

\item{check_for_new_data}{Logical. Whether to skip checking for updates (to save time)}

\item{force_redownload_all}{}
}
\description{
Check for, and download MSDS data
}
