% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msds_download_data.R
\name{msds_download_data}
\alias{msds_download_data}
\title{Check for, and download MSDS data}
\usage{
msds_download_data(destination = "data/msds_download", relative_path = TRUE)
}
\arguments{
\item{destination}{Character. The filepath where downloads will be saved}

\item{relative_path}{Logical (default TRUE). Whether the desination path will be treated as relative to working directory}
}
\description{
Check for, and download MSDS data
}
\examples{
# download data to the default directory path
msds_download_data()

# download data to a different (relative) directory path
msds_download_data(destination = "a/different/directory")

# download data to an absolute directory path, or network path
msds_download_data(destination = "C:/your/hard/drive/msds_data", relative_path = FALSE)
}
