% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msds_tidy_dq.R
\name{msds_tidy_dq}
\alias{msds_tidy_dq}
\title{Combine and tidy data quality data}
\usage{
msds_tidy_dq(data_path = "data/msds_download")
}
\arguments{
\item{data_path}{Character string defining the path to the parent data folder}
}
\value{
A dataframe of data quality data
}
\description{
Combine and tidy data quality data
}
