% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msds_tidy_measures.R
\name{msds_tidy_measures}
\alias{msds_tidy_measures}
\title{Combine and tidy measures data files}
\usage{
msds_tidy_measures(data_path = "data/msds_download", relative_path = TRUE)
}
\arguments{
\item{data_path}{Character string defining the path to the parent data folder}

\item{relative_path}{Logical (default TRUE). Whether the data_path will be treated as relative to working directory}
}
\value{
A dataframe of measures data
}
\description{
Combine and tidy measures data files
}
